perf(Background shading, camera): Moved background shading calculation to server side, fine tuned camera range settings
Different versions of backgrounds are calculated serverside during the build process. Game will load faster since the calculation doesn't need to be done on the client after every load. For development, might be a bit slower because all backgrounds get calculated on every reload instead of just the current background. Will improve later.

Fine tuned ability to control the default look position of the camera and how much it can adjust horizontally and vertically with the mouse.