feat(Mesh changing): Added mesh changing, fixed bugs related to advance action
Created command queue which allows dialogue to send a command to have an entity in the scene change mesh to an alternate mesh. It can also be used for temporary changes to a location that don't need to be saved to an actual check.

Changed fading effect to use three.js's animation system.

Fixed bug where using keyboard to advance doesn't work after clicking on html ui.
Fixed bug where the advance action doesn't change if the mouse doesn't move.