feat(Carriage, rendering coordinate flexibility, camera default position): Added Carriage location, improved rendering coordinate flexibility, removed camera default position dependency
Removed constant unit and depth for locations. These values are now variable and are set per level within the location file.

I used the above changes to build the Carriage location.

Removed camera default position as a dependency for loadEntities to eliminate a race condition when a new location is loaded.